Move limit types to shared folder
